requirements.txt
flask
gizeh
gunicorn
huesdk
imageio
matplotlib
moviepy
pyyaml
webcolors
wheel